+++ src/Observers/FileObserver.php
@@ -69,7 +69,9 @@ class FileObserver
             $file->user_deleted_id = auth()->user()->id ?? null;
             $file->saveQuietly();
 
-            $file->activity->delete();
+            if($file->activity) {
+                $file->activity->delete();
+            }
         }
     }
