+++ src/Observers/TaskObserver.php
@@ -69,7 +69,9 @@ class TaskObserver
             $task->user_deleted_id = auth()->user()->id ?? null;
             $task->saveQuietly();
 
-            $task->activity->delete();
+            if($task->activity) {
+                $task->activity->delete(); 
+            }
         }
     }
